requirements.txt
openai-whisper
torch
torchaudio
transformers
numpy
pandas
moviepy
